requirements.txt
# 开发基本要求
# 不过我个人建议还是使用PDM安装依赖，requirements仅作为开发安装依赖
## Graia
graia-amnesia==0.7.1
graia-ariadne<1.0.0,>=0.10.2
graia-broadcast==0.19.1
graia-saya==0.0.17
graia-scheduler==0.1.5
graiax-fastapi==0.2.1
creart-graia==0.1.5
# Tips:目前Python3.12未支持Graia组件
## 网络要求
requests<3.0.0,>=2.28.1
aiohttp<4.0.0,>=3.8.3
requests-cache<1.0.0,>=0.9.7
## 数据库要求
redis<5.0.0,>=4.4.0
aiomysql<1.0.0,>=0.1.1
pymysql<2.0.0,>=1.0.2
#pysqlite3 # 这是标准库，如果IDE报错请把这串注释去除
## 分词工具
jieba<1.0.0,>=0.42.1
numpy>=1.24.2
# YMB配置
pyyaml<7.0,>=6.0
pyjson
# 腾讯云安全
tencentcloud-sdk-python-tms>=3.0.967
tencentcloud-sdk-python-ims>=3.0.967
# 剩下有的没的
pathlib>=1.0.1
opencc-python-reimplemented<1.0.0,>=0.1.6
portalocker>=2.7.0
rich>=13.4.2
lib-not-dr>=0.2.2
# 打包工具（不过应该没有人专门给一个QQBot打包成exe罢，我就注释掉了，需要自行解除注释）
# nuitka